CEDS-1193: moved test around so we are able to see what work is needed to cover different types of testing. - Added integration test for connector
