+++ test/integration/uk/gov/hmrc/exports/base/IntegrationTestSpec.scala
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.exports.models
+package integration.uk.gov.hmrc.exports.base
 
-import uk.gov.hmrc.auth.core.retrieve.{Credentials, Name}
-import uk.gov.hmrc.auth.core.{AffinityGroup, Enrolments}
+import com.google.inject.AbstractModule
+import org.scalatest.concurrent.Eventually
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
+import play.api.inject.guice.GuiceableModule
+import uk.gov.hmrc.play.test.UnitSpec
 
-case class CustomsDeclarationsResponse(status: Int, conversationId: Option[String] = None)
+case class IntegrationTestModule() extends AbstractModule {
+  def configure(): Unit = ()
 
-case class SignedInUser(
-  credentials: Credentials,
-  name: Name,
-  email: Option[String],
-  eori: String,
-  externalId: String,
-  internalId: Option[String],
-  affinityGroup: Option[AffinityGroup],
-  enrolments: Enrolments
-)
+  def asGuiceableModule: GuiceableModule = GuiceableModule.guiceable(this)
+}
+
+trait IntegrationTestSpec extends UnitSpec
+  with BeforeAndAfterEach with BeforeAndAfterAll with Eventually
